Fix broken test
The original estimate of 4-5 iterations of #9 was based on observations
on a machine using 6 worker processes. Now, using 10 processes (and a
different error tolerance; I don't recall the original one) it looks
like convergence depends more on the time span of the fine solution
rather than the bare number of iterations. As we only care for "early
convergence" to occur at some point (`maxiters < length(workers)`),
increasing `maxiters` will suffice.