Add support for threads
Hook into `DiffEqBase.solve` instead of `DiffEqBase.__solve`. This will
skip all tests of the input such as the danger of an integer time span.
It was necessary however as some of those tests demand the user to
specify a time step as this algorithm isn't adaptive. Since the concept
of adaptivity doesn't translate (easily) to parallel-in-time algorithms,
I've decided to skip the tests altogether.