Refactor cancellation
Instead of having a single "lifeline" to all of the workers that gets
cut on cancellation (by closing a `RemoteChannel`), send a cancellation
signal on the same channel that is used for new initial values. While
making the workers behave more like a state machine, this needs no extra
tasks to concurrently check for cancellation and synchronize with the
task performing the actual computations.

Furthermore, this removes the dead end mentioned in !12, i.e. the point
where the workers wait for either new data or cancellation.