Fix detection of failures on workers
Waiting on a `Future` (i.e. some operation running remotely), as opposed
to waiting on a `Task`, doesn't throw an exception when that operation
failed.  Hence, replace all `try ... catch` blocks by checks of the type
of object returned by `fetch`. On failure, it is an `Exception` object
and `nothing` otherwise.

Furthermore, intercept exceptions on workers/stages and send a status
update. This way, workers are the only ones sending those. The event
handler running on the managing process (or user code, likely running on
the managing process as well) are the only ones who might cancel the
pipeline.