Fix iteration counts
Previously, the corner case of reaching the maximum number of iterations
wasn't handled properly. It is only safe to send the fine solution as
the converged solution if the maximum number of iterations hasn't been
reached yet (k < K) and the previous stage converged (didconverge(msg)).
The latter is also the case if the fine integrator reached the current
time slice (k == n).