Refactor for structured concurrency
i.e. have every function outlive all the tasks it spawns, i.e. prevent
functions from leaking tasks (wherever possible).

https://en.wikipedia.org/wiki/Structured_concurrency

Pipeline cancellation is one notable exception, currently. I don't have
a clear path in mind on how to fix this. If I (or any other developer)
makes a "stupid mistake" such as using an undefined method in a part of
the code that is run asynchronously (e.g. `remotecall_foo` to spawn the
workers, or somewhere inside `fsolve` and `csolve`), the initial value
might never be consumed from the first connection. As all connections
only have a buffer size of 1, this will cause `cancel_pipeline!` to
block.
One could increase the buffer size, but that would cause the stages to
get out of sync too much, increasing memory consumption and merely
deferring the issue.
However, in the situations described above, a leaking task is only a
symptom of a bigger problem, so this might be a non-issue.