Fix deadlock due to convergence of later stages
If a stage late in the pipeline converges, it won't consume any values
sent from the previous stage. If the previous stage performs 2
iterations more than the current stage, this used to cause a deadlock
as the send will never complete. Resolve this issue by increasing the
buffer size of the connections.

This solution is not ideal, since the earlier stages might have drifted
off causing the "convergence" of the later stage to be invalid.