Add LazyFormatLogger
that opens a logfile upon arrival of the first log message. This is one
way around the issue that one cannot send an open file handle (IOStream)
to a worker process. Before the "first use", this logger is safe to be
sent over the network.

Using one of these per worker stage should be preferred over
CommunicatingLogger and CommunicatingObserver, as it better preserves
information over the event of a node failure or deadlock.