Don't remove Stage from container
If a stage fails hard, or if due to some other error the StageRef
container holds another Stage, it may otherwise be impossible to get the
Stage back and retrieve further information from it. This will be
helpful in the investigation of errors and allows to write the last
computed values to disk regardless of a failure.