Fix early convergence
If the previous stage has converged, while the current one hasn't,
it is either `k == n`, i.e. the stages occur "early" in the pipeline,
or `k < n`, i.e. the stages converged due to the parareal method.
This latter is what is referred to as "early convergence" by some
comments in the code.

This hasn't been reflected properly in the code.